feat: add lunr search to navbar (#105)
**Motivation**
When going through the docs it requires to open the categories or scroll
within the documents to find something specific which takes extra time.
Adding the search functionality allows to navigate faster within the
docs.

**PR Changes**
- adds [lunar.js](https://lunrjs.com/) search to the navigation bar


**Screenshot**

![image](https://github.com/user-attachments/assets/60572ff3-ec87-4b7b-9fbd-7a831d09a35e)


**Note**
It works only in prod build so I've been doing `npm run build` and then
`http-server build` to serve the production version of docs to test it
out