LICENSE as CC-BY-SA 4.0 (#137)
After legal review it was determined we want to use the Creative Commons
Attribution 4.0 license. This updates the license file from the template
example here:
https://github.com/jslicense/CC-BY-SA-4.0/blob/main/CC-BY-SA-4.0

Once approved and merged, the current plan is to remove the prior
LICENSE commit through a force push of a squashed commit, this will
prevent having the prior history of the docs being incorrectly licensed.