chore(docs): DSPX-1442 update Java & TS to match Go-lang (#133)
The rendered page for **Creating TDF's** , with the heading "Encrypting
/ Decrypting zTDF" is here:

* https://opentdf.io/sdks/tdf

You'll notice, on that page, that the Go example is much more
approachable and provides friendly output to the user about the status
of work being performed.

This PR is meant to update the Java & TypeScript code sample to be just
as approachable.